feat: Put blob support, azureite local testing support
The new Put interface on the storer uses the azure PUT interface (via
the sdk Upload method). This has 'create or replace' semantics and
allows the metadata and tags to be set in the same operation.

These semantics are require for forestrie. For blobs that are small
enough to be uploaded in one shot (less than 10's of megabytes) this is
also significantly more performant.

Additionaly, adds support for testing localy using the azurite
opensource blob, queue & tablel emulator. To take advantage of this:
* Use the taskfile support to launch the emulator in a docker container
* In go-lang unit tests, use NewDev to create your storer instance. Once
  created it behaves mostly as the cloud service. Though there are some
  restrictions on tag and meta data sizes that are smaller than the
  production versions.

AB#8656