feat: etag support for put & read blob
The introduction of ETag support makes it possible to do two things:

1. read/modify/write concurrency guards, so that values are written back
   only if the destination content hasn't been modified since being
   read.
2. Efficient cache refreshing without fetching the blob data. still
   requires request round trip, but no data is returned in the response
   if the content is unchanged since the last read.

1. is going to give us crash fault tollerance and race reconciliaiton
   for updating forestrie blobs.

behaviour change:

Previously Read swallowed any error that wasn't accompanied by an io.EOF
condition. This causes the return states to be very confusing when
dealint with If- header responses. The err is now returned in that case.

The arrangements to ensure body.Read is called remain as before

AB#8656